Admin level 1 africa seasonal parameters

These datasets represent the data fitted within the Imperial College Malaria model for relating seasonal profiles to malaria transmission intensity at level 1 admin regions across Africa

admin_units_seasonal

Format

A dataframe of 576 observations of 13 variables:

$admin_units_seasonal: A dataframe of admin units and their seasonal parameters

country: Country string

admin1: Admin 1 string

map_prev_2010: 2010 Atlas map microscopy prevalence in 2-10 year olds

id: Numeric vector of 1:576

ID_1: Numeric vector referencing shape file geoshape ids

a0: Average value of fourier series

a1: First of partial series cos terms

b1: First of partial series sine terms

a2: Second of partial series cos terms

b2: Second of partial series sine terms

a3: Third of partial series cos terms

b3: Third of partial series sine terms

theta_c: Rainfall normalising constant